Present the whois results as table
